Implementando as exceptions e adaptando o código para seus usos.

+++ src/main/java/br/com/fiap/diamondgames/service/JogadorService.java
@@ -1,5 +1,6 @@
 package br.com.fiap.diamondgames.service;
 
+import br.com.fiap.diamondgames.exception.ResourceNotFoundException;
 import br.com.fiap.diamondgames.model.Jogador;
 import br.com.fiap.diamondgames.projection.JogadorResumoProjection;
 import br.com.fiap.diamondgames.repository.JogadorRepository;
@@ -7,8 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 public class JogadorService {
@@ -20,8 +19,9 @@ public class JogadorService {
         return repository.findAllResumo(pageable);
     }
 
-    public Optional<Jogador> buscarPorId(Long id) {
-        return repository.findById(id);
+    public Jogador buscarPorId(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Jogador não encontrado com o ID: " + id));
     }
 
     public Jogador salvar(Jogador jogador) {
@@ -29,6 +29,7 @@ public class JogadorService {
     }
 
     public void deletar(Long id) {
+        if (!existe(id)) throw new ResourceNotFoundException("Jogador não encontrado para deleção. ID: " + id);
         repository.deleteById(id);
     }
